requirements.txt
flask-wtf
email-validator
flask-bootstrap